perf(sqlite): journal appended artifact parts, so a streaming append stops paying for the stream so far
The incremental artifact path saved the Rust-side serialization of the whole
task and not the database-side one: `json_set` parses and rewrites the entire
JSON document, so the statement stayed linear in what the stream had already
produced. That was recorded in the code as a known limit, with the reasoning
that the per-event round trip dominated anyway.

Measured across document sizes rather than at one, the reasoning does not
hold — the round trip is a constant and the rewrite is not. Same probe
through `SqliteTaskStore`, release build, 200 appends per arm, best of three,
run against the previous commit and this one:

  parts   json_set delta   vs save      journalled   vs save
     10          171.6us     0.71x         112.1us     0.26x
    100          289.5us     0.72x         140.1us     0.24x
    300          586.8us     0.98x         123.9us     0.17x
   1000         1574.8us     1.39x         137.2us     0.10x
   3000         4516.1us     1.48x         124.7us     0.03x

The middle column is worse than "fails to help". Past roughly 300 parts the
incremental path was *slower than rewriting the whole record* — it paid a
`json_set` parse of the document on top of everything a full save does. An
optimisation named for being incremental was a pessimisation for exactly the
long streams it existed for.

Appended parts now go to their own append-only rows and the document is left
alone, which makes the append flat: ~130us whether the artifact holds ten
parts or three thousand. At 3,000 parts, 4.5ms becomes 0.12ms.

The document stays the record of the task. The table holds only parts
appended since it was last written whole, every read splices them back, and a
full `save` rewrites the document and clears them in one transaction — so a
part is in the document or in the journal, never both and never neither. That
also bounds the table without a compactor: every status change persists the
task whole, so the journal holds at most one status interval, and a finished
task has none.

`seq` is the part's index rather than a counter, which is what makes an
append need no read to place itself, unable to race another into the same
slot, and idempotent on replay.

This costs `save` something, and the number is in the table above: it now
opens a transaction and clears the journal as well as writing the document,
moving a ten-part save from ~241us to ~427us. Recorded rather than buried —
it is the right trade only because the two run at different rates, a handful
of saves per task against thousands of appends.

Four injections. One confirmed a test; three found that tests were missing:

  * `list` not splicing — caught. It was the read path most easily
    forgotten, since every existing equivalence test goes through `get`.
  * `save` not clearing the journal — **not** caught, and the reason matters:
    `splice` skips positions the document already holds, so stale rows return
    correct tasks forever while the table grows without limit. Reading tasks
    back cannot see it. `a_full_save_empties_the_journal` asserts on the
    table instead, and now catches it.
  * `delete` leaving rows behind — not caught, because `save` clearing the
    journal already defends id reuse. The explicit delete still matters for a
    task deleted and never re-saved, so it stays.
  * splice not skipping parts already in the document — caught, though a
    `head -6` on the first run hid the failure and briefly suggested
    otherwise.

Also splits the file it grew: `sqlite_store.rs` was 1,529 lines and would
have been 1,797. It is now `sqlite_store/` at 714, with the two test modules
and the journal in their own files, and the baseline entry is renamed rather
than added. `--update` strips every comment from that file, so the note it
removed is restored along with one saying so.
