+++ bindings/a2a-protocol-slimrpc/src/server/methods/mod.rs
@@ -430,91 +430,4 @@ fn single_response_stream(response: pb::SendMessageResponse) -> EventStream {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-
-    fn metadata(pairs: &[(&str, &str)]) -> HashMap<String, String> {
-        pairs
-            .iter()
-            .map(|(k, v)| ((*k).to_string(), (*v).to_string()))
-            .collect()
-    }
-
-    /// The gap this closes: SLIMRPC's routing and timing keys arrive in the
-    /// same flat map as the caller's A2A headers, and used to be handed to the
-    /// handler intact. `service` and `method` are ordinary enough words that a
-    /// `HeaderTenantResolver` keyed on either would have resolved a tenant from
-    /// transport plumbing.
-    #[test]
-    fn transport_keys_never_reach_the_handler() {
-        let headers = a2a_headers(metadata(&[
-            (slim_rpc::DEADLINE_KEY, "2026-01-01T00:00:00Z"),
-            (slim_rpc::STATUS_CODE_KEY, "0"),
-            ("rpc-id", "17"),
-            ("service", "lf.a2a.v1.A2AService"),
-            ("method", "SendMessage"),
-            ("authorization", "Bearer token"),
-            ("x-tenant-id", "acme"),
-        ]))
-        .expect("no version declared, which is accepted");
-
-        assert_eq!(
-            headers,
-            metadata(&[("authorization", "Bearer token"), ("x-tenant-id", "acme")]),
-            "only the caller's own headers survive"
-        );
-    }
-
-    /// SLIM does not promise a case for its keys, and a filter that missed
-    /// `Service` while catching `service` would leak exactly the case an
-    /// attacker would pick.
-    #[test]
-    fn transport_keys_are_filtered_regardless_of_case() {
-        let headers = a2a_headers(metadata(&[
-            ("RPC-ID", "17"),
-            ("Service", "x"),
-            ("METHOD", "y"),
-            ("x-tenant-id", "acme"),
-        ]))
-        .expect("must succeed");
-
-        assert_eq!(headers, metadata(&[("x-tenant-id", "acme")]));
-    }
-
-    /// §3 requires the version to be transmitted; a value this server does not
-    /// implement must be refused rather than treated as absent.
-    #[test]
-    fn an_unsupported_version_is_rejected() {
-        let err = a2a_headers(metadata(&[("a2a-version", "0.3")]))
-            .expect_err("0.3 is not a version this server speaks");
-
-        assert!(
-            format!("{err:?}").contains("0.3"),
-            "the rejection should name the version, got: {err:?}"
-        );
-    }
-
-    #[test]
-    fn a_supported_version_passes_through_to_the_handler() {
-        let headers =
-            a2a_headers(metadata(&[("a2a-version", "1.0")])).expect("1.0 is what we implement");
-
-        assert_eq!(
-            headers.get("a2a-version").map(String::as_str),
-            Some("1.0"),
-            "the version is a service parameter, not a key to consume"
-        );
-    }
-
-    /// The deliberate leniency, pinned so it cannot be tightened by accident:
-    /// the official `a2a-slimrpc` crate sends no version, and rejecting an
-    /// absent one would refuse every call from the A2A project's own Rust SDK.
-    #[test]
-    fn an_absent_version_is_accepted_for_interop() {
-        let headers = a2a_headers(metadata(&[("authorization", "Bearer t")]))
-            .expect("an unversioned caller is still served");
-
-        assert_eq!(headers, metadata(&[("authorization", "Bearer t")]));
-    }
-}
+mod tests;
